requirements.txt
Django==1.9.2
django-filter==0.12.0
djangorestframework==3.3.2
facebook-sdk==0.4.0
Markdown==2.6.5
MySQL-python==1.2.5
requests==2.9.1
wheel==0.24.0
